Задание 5.14. Модуляция
